+++ KlxPiaoAPI/FileUtils.cs
@@ -13,7 +13,7 @@ namespace KlxPiaoAPI
         /// 加载指定路径或资源名称的字体。
         /// </summary>
         /// <param name="pathOrResourceName">字体文件的路径或资源名称。</param>
-        /// <param name="isResource">指示是否从资源加载字体。如果为 <<c>true</c>>，则从嵌入的资源加载字体；否则从文件路径加载字体。</param>
+        /// <param name="isResource">指示是否从资源加载字体。如果为 true，则从嵌入的资源加载字体；否则从文件路径加载字体。</param>
         /// <returns>返回的 <see cref="FontFamily"/> 对象。</returns>
         /// <exception cref="Exception">指定的文件路径或资源不存在时抛出。</exception>
         public static FontFamily LoadFontFamily(string pathOrResourceName, bool isResource = false)